Split "new" miri crate into modules
